refactor: rename to RNMLKitBarcodeScanning
