+++ modules/react-native-mlkit-barcode-scanning/android/src/main/java/expo/modules/mlkitbarcodescanning/ReactNativeMlkitBarcodeScanningModule.kt
@@ -4,15 +4,15 @@ import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import java.net.URL
 
-class ReactNativeMlkitBarcodeScanningModule : Module() {
+class RNMLKitBarcodeScanningModule : Module() {
   // Each module class must implement the definition function. The definition consists of components
   // that describes the module's functionality and behavior.
   // See https://docs.expo.dev/modules/module-api for more details about available components.
   override fun definition() = ModuleDefinition {
     // Sets the name of the module that JavaScript code will use to refer to the module. Takes a string as an argument.
     // Can be inferred from module's class name, but it's recommended to set it explicitly for clarity.
-    // The module will be accessible from `requireNativeModule('ReactNativeMlkitBarcodeScanning')` in JavaScript.
-    Name("ReactNativeMlkitBarcodeScanning")
+    // The module will be accessible from `requireNativeModule('RNMLKitBarcodeScanning')` in JavaScript.
+    Name("RNMLKitBarcodeScanning")
 
     // Defines constant property on the module.
     Constant("PI") {
@@ -38,9 +38,9 @@ class ReactNativeMlkitBarcodeScanningModule : Module() {
 
     // Enables the module to be used as a native view. Definition components that are accepted as part of
     // the view definition: Prop, Events.
-    View(ReactNativeMlkitBarcodeScanningView::class) {
+    View(RNMLKitBarcodeScanningView::class) {
       // Defines a setter for the `url` prop.
-      Prop("url") { view: ReactNativeMlkitBarcodeScanningView, url: URL ->
+      Prop("url") { view: RNMLKitBarcodeScanningView, url: URL ->
         view.webView.loadUrl(url.toString())
       }
       // Defines an event that the view can send to JavaScript.
