+++ modules/react-native-mlkit-barcode-scanning/android/src/main/java/expo/modules/mlkitbarcodescanning/ReactNativeMlkitBarcodeScanningView.kt
@@ -7,7 +7,7 @@ import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoView
 
-class ReactNativeMlkitBarcodeScanningView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
+class RNMLKitBarcodeScanningView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
   // Creates and initializes an event dispatcher for the `onLoad` event.
   // The name of the event is inferred from the value and needs to match the event name defined in the module.
   private val onLoad by EventDispatcher()
