+++ modules/react-native-mlkit-barcode-scanning/index.ts
@@ -1,4 +1,4 @@
-// Reexport the native module. On web, it will be resolved to ReactNativeMlkitBarcodeScanningModule.web.ts
-// and on native platforms to ReactNativeMlkitBarcodeScanningModule.ts
-export { default } from './src/ReactNativeMlkitBarcodeScanningModule';
-export * from  './src/ReactNativeMlkitBarcodeScanning.types';
+// Reexport the native module. On web, it will be resolved to RNMLKitBarcodeScanningModule.web.ts
+// and on native platforms to RNMLKitBarcodeScanningModule.ts
+export { default } from './src/RNMLKitBarcodeScanningModule';
+export * from  './src/RNMLKitBarcodeScanning.types';
