+++ modules/react-native-mlkit-barcode-scanning/ios/ReactNativeMlkitBarcodeScanning.podspec
@@ -1,5 +1,5 @@
 Pod::Spec.new do |s|
-  s.name           = 'ReactNativeMlkitBarcodeScanning'
+  s.name           = 'RNMLKitBarcodeScanning'
   s.version        = '1.0.0'
   s.summary        = 'A sample project summary'
   s.description    = 'A sample project description'
