+++ modules/react-native-mlkit-barcode-scanning/ios/ReactNativeMlkitBarcodeScanningModule.swift
@@ -1,14 +1,14 @@
 import ExpoModulesCore
 
-public class ReactNativeMlkitBarcodeScanningModule: Module {
+public class RNMLKitBarcodeScanningModule: Module {
   // Each module class must implement the definition function. The definition consists of components
   // that describes the module's functionality and behavior.
   // See https://docs.expo.dev/modules/module-api for more details about available components.
   public func definition() -> ModuleDefinition {
     // Sets the name of the module that JavaScript code will use to refer to the module. Takes a string as an argument.
     // Can be inferred from module's class name, but it's recommended to set it explicitly for clarity.
-    // The module will be accessible from `requireNativeModule('ReactNativeMlkitBarcodeScanning')` in JavaScript.
-    Name("ReactNativeMlkitBarcodeScanning")
+    // The module will be accessible from `requireNativeModule('RNMLKitBarcodeScanning')` in JavaScript.
+    Name("RNMLKitBarcodeScanning")
 
     // Defines constant property on the module.
     Constant("PI") {
@@ -34,9 +34,9 @@ public class ReactNativeMlkitBarcodeScanningModule: Module {
 
     // Enables the module to be used as a native view. Definition components that are accepted as part of the
     // view definition: Prop, Events.
-    View(ReactNativeMlkitBarcodeScanningView.self) {
+    View(RNMLKitBarcodeScanningView.self) {
       // Defines a setter for the `url` prop.
-      Prop("url") { (view: ReactNativeMlkitBarcodeScanningView, url: URL) in
+      Prop("url") { (view: RNMLKitBarcodeScanningView, url: URL) in
         if view.webView.url != url {
           view.webView.load(URLRequest(url: url))
         }
