+++ modules/react-native-mlkit-barcode-scanning/src/ReactNativeMlkitBarcodeScanningModule.ts
@@ -1,10 +1,10 @@
 import { NativeModule, requireNativeModule } from 'expo';
 
-import { BarcodeScannerResult } from './ReactNativeMlkitBarcodeScanning.types';
+import { BarcodeScannerResult } from './RNMLKitBarcodeScanning.types';
 
-declare class ReactNativeMlkitBarcodeScanningModule extends NativeModule {
+declare class RNMLKitBarcodeScanningModule extends NativeModule {
   process(imagePath: string): Promise<BarcodeScannerResult>;
 }
 
 // This call loads the native module object from the JSI.
-export default requireNativeModule<ReactNativeMlkitBarcodeScanningModule>('ReactNativeMlkitBarcodeScanning');
+export default requireNativeModule<RNMLKitBarcodeScanningModule>('RNMLKitBarcodeScanning');
